+++ packages/web/src/app/doc-pages/components/modal-doc/modal-data.ts
@@ -66,6 +66,11 @@ const modalData = {
       type: '() => void',
       description: 'Callback for every time the modal is being closed.',
     },
+    maxWidth: {
+      isRequired: false,
+      type: 'string | undefined',
+      description: 'Overwrite the max width of the modal (e.g. "100px", "90%").',
+    },
   },
   package: 'npm install @elvia/elvis-modal',
   codeImportReact: `import { Modal } from '@elvia/elvis-modal/react';`,