+++ packages/web/src/app/shared/component-documentation/component-example/component-example-generator/ceg-filters/ceg-filters.component.ts
@@ -25,9 +25,11 @@ export class CegFiltersComponent implements OnInit {
   @Output() hasVisibleFilters = new EventEmitter();
   codeAngularSub: Subscription;
   codeReactSub: Subscription;
+  codeVueSub: Subscription;
   codeNativeSub: Subscription;
   codeReact;
   codeAngular;
+  codeVue;
   codeNative;
 
   counterNumber: number;
@@ -41,11 +43,15 @@ export class CegFiltersComponent implements OnInit {
     this.codeReact = this.componentData.codeReact;
     this.codeAngular = this.componentData.codeAngular;
     this.codeNative = this.componentData.codeNativeHTML;
+    this.codeVue = this.componentData.codeVue ? this.componentData.codeVue : '';
     this.codeAngularSub = this.cegService.listenCodeAngular().subscribe((code: string) => {
       this.codeAngular = code;
     });
     this.codeReactSub = this.cegService.listenCodeReact().subscribe((code: string) => {
       this.codeReact = code;
+    });
+    this.codeVueSub = this.cegService.listenCodeVue().subscribe((code: string) => {
+      this.codeVue = code;
     });
     this.codeNativeSub = this.cegService.listenCodeNative().subscribe((code: string) => {
       this.codeNative = code;
@@ -60,6 +66,9 @@ export class CegFiltersComponent implements OnInit {
     }
     if (this.codeReactSub) {
       this.codeReactSub.unsubscribe();
+    }
+    if (this.codeVue) {
+      this.codeVueSub.unsubscribe();
     }
     if (this.codeNativeSub) {
       this.codeNativeSub.unsubscribe();
@@ -149,6 +158,7 @@ export class CegFiltersComponent implements OnInit {
   updateNewCode(): void {
     this.cegService.updateCodeReact(this.codeReact);
     this.cegService.updateCodeAngular(this.codeAngular);
+    this.cegService.updateCodeVue(this.codeVue);
     this.cegService.updateCodeNative(this.codeNative);
   }
 
@@ -157,18 +167,21 @@ export class CegFiltersComponent implements OnInit {
     const elNameW = this.componentData.elementNameW;
     this.codeReact = this.cegService.addNewProp(this.codeReact, attr, newValue, 'react', type, elNameR);
     this.codeAngular = this.cegService.addNewProp(this.codeAngular, attr, newValue, 'angular', type, elNameW);
+    this.codeVue = this.cegService.addNewProp(this.codeVue, attr, newValue, 'vue', type, elNameW);
     this.codeNative = this.cegService.addNewProp(this.codeNative, attr, newValue, 'native', type, elNameW);
   }
 
   replaceOldProps(attr: string, newValue: string, type: string): void {
     this.codeReact = this.cegService.replaceOldProp(this.codeReact, attr, newValue, 'react', type);
     this.codeAngular = this.cegService.replaceOldProp(this.codeAngular, attr, newValue, 'angular', type);
+    this.codeVue = this.cegService.replaceOldProp(this.codeVue, attr, newValue, 'vue', type);
     this.codeNative = this.cegService.replaceOldProp(this.codeNative, attr, newValue, 'native', type);
   }
 
   removeProps(attr: string): void {
     this.codeReact = this.cegService.removeProp(this.codeReact, attr);
     this.codeAngular = this.cegService.removeProp(this.codeAngular, attr);
+    this.codeVue = this.cegService.removeProp(this.codeVue, attr);
     this.codeNative = this.cegService.removeProp(this.codeNative, attr);
   }
 
@@ -177,21 +190,17 @@ export class CegFiltersComponent implements OnInit {
     const elNameW = this.componentData.elementNameW;
     this.codeReact = this.cegService.addNewSlotAndProp(this.codeReact, attr, value, 'react', elNameR);
     this.codeAngular = this.cegService.addNewSlotAndProp(this.codeAngular, attr, value, 'angular', elNameW);
+    this.codeVue = this.cegService.addNewSlotAndProp(this.codeVue, attr, value, 'vue', elNameW);
     this.codeNative = this.cegService.addNewSlotAndProp(this.codeNative, attr, value, 'native', elNameW);
   }
 
-  removeSlotAndProp(attr: string, oldValue: string): void {
+  removeSlotAndProp(attr: string, value: string): void {
     const elNameR = this.componentData.elementNameR;
     const elNameW = this.componentData.elementNameW;
-    this.codeReact = this.cegService.removeSlotAndProp(this.codeReact, attr, oldValue, 'react', elNameR);
-    this.codeAngular = this.cegService.removeSlotAndProp(
-      this.codeAngular,
-      attr,
-      oldValue,
-      'angular',
-      elNameW,
-    );
-    this.codeNative = this.cegService.removeSlotAndProp(this.codeNative, attr, oldValue, 'native', elNameW);
+    this.codeReact = this.cegService.removeSlotAndProp(this.codeReact, attr, value, 'react', elNameR);
+    this.codeAngular = this.cegService.removeSlotAndProp(this.codeAngular, attr, value, 'angular', elNameW);
+    this.codeVue = this.cegService.removeSlotAndProp(this.codeVue, attr, value, 'vue', elNameW);
+    this.codeNative = this.cegService.removeSlotAndProp(this.codeNative, attr, value, 'native', elNameW);
   }
 
   updateRadioProp(attr: string, newValue: string, type: string): void {