+++ packages/components/components/elvis-component-wrapper/CHANGELOG.md
@@ -1,4 +1,16 @@
-# Elvia Accordion Changelog
+# Elvia Component Wrapper Changelog
+
+## 2.1.5 (22.11.21)
+
+### Bug fix
+
+- Now allowing props to have value null
+
+## 2.1.4 (17.11.21)
+
+### Bug fix
+
+- Improved conditional elements type function
 
 ## 2.1.4 (17.11.21)